chore: move vega entities out of receipts directory
