+++ frontend/components/vega-entities/market-link.tsx
@@ -2,7 +2,7 @@ import { truncateMiddle } from '@vegaprotocol/ui-toolkit'
 
 import config from '!/config'
 
-import { ExternalLink } from '../../../external-link'
+import { ExternalLink } from '../external-link'
 
 export const locators = {
   marketLink: 'market-link'
