+++ frontend/components/vega-entities/node-link.tsx
