+++ frontend/components/vega-entities/proposal-link.tsx
